Crate tuxedo_template_runtime

The Tuxedo Template Runtime is an example runtime that uses most of the pieces provided in the wardrobe.

Runtime developers wishing to get started with Tuxedo should consider copying this template.

Re-exports

pub use amoeba;

pub use kitties;

pub use money;

pub use poe;

pub use runtime_upgrade;

pub use timestamp;

Modules

api

genesis

Helper module to build a genesis configuration for the template runtime.

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

Runtime

The main struct in this module.

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

Enums

OuterConstraintChecker

A constraint checker is a piece of logic that can be used to check a transaction. For any given Tuxedo runtime there is a finite set of such constraint checkers. For example, this may check that input token values exceed output token values.

OuterConstraintCheckerError

This type is generated by the #[tuxedo_constraint_checker] macro. It is a combined error type for the errors of each individual checker.

OuterVerifier

A verifier checks that an individual input can be consumed. For example that it is signed properly To begin playing, we will have two kinds. A simple signature check, and an anyone-can-consume check.

OuterVerifierRedeemer

This type is generated by the #[tuxedo_verifier] macro. It is a combined redeemer type for the redeemers of each individual verifier.

Constants

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Block

Executive

Output

Transaction