Type Alias tuxedo_template_runtime::Transaction

pub type Transaction = Transaction<OuterVerifier, OuterConstraintChecker>;

Aliased Type

struct Transaction {
    pub inputs: Vec<Input>,
    pub peeks: Vec<OutputRef>,
    pub outputs: Vec<Output<OuterVerifier>>,
    pub checker: OuterConstraintChecker,
}

Fields
inputs: Vec<Input>

Existing pieces of state to be read and consumed from storage

peeks: Vec<OutputRef>

Existing state to be read, but not consumed, from storage

outputs: Vec<Output<OuterVerifier>>

New state to be placed into storage

checker: OuterConstraintChecker

Which piece of constraint checking logic is used to determine whether this transaction is valid