Crate tuxedo_template_runtime

Modules

amoeba

This file represents a simple example Tuxedo piece that tracks amoeba populations. Amoeba’s can be affected in three ways throughout their lifecycle.

api

kitties

An NFT game inspired by cryptokitties. This is a game which allows for kitties to be bred based on a few factors 1.) Mom and Tired have to be in a state where they are ready to breed 2.) Each Mom and Dad have some DNA and the child will have unique DNA combined from the both of them Linkable back to the Mom and Dad 3.) The game also allows Kitties to have a cooling off period inbetween breeding before they can be bred again. 4.) A rest operation allows for a Mom Kitty and a Dad Kitty to be cooled off

money

An implementation of a simple fungible token.

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

GenesisConfig

Runtime

The main struct in this module.

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

Enums

OuterConstraintChecker

A constraint checker is a piece of logic that can be used to check a transaction. For any given Tuxedo runtime there is a finite set of such constraint checkers. For example, this may check that input token values exceed output token values.

OuterConstraintCheckerError

This type is generated by the #[tuxedo_constraint_checker] macro. It is a combined error type for the errors of each individual checker.

OuterVerifier

A verifier checks that an individual input can be consumed. For example that it is signed properly To begin playing, we will have two kinds. A simple signature check, and an anyone-can-consume check.

Constants

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

Block

BlockNumber

Executive

Header

Output

Transaction